Refactor forward search
