+++ src/nl/rubensten/texifyidea/action/sumatra/ConfigureInverseSearchAction.kt
@@ -6,7 +6,7 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.application.ApplicationNamesInfo
 import com.intellij.openapi.application.PathManager
 import com.intellij.openapi.ui.DialogBuilder
-import nl.rubensten.texifyidea.run.isSumatraAvailable
+import nl.rubensten.texifyidea.run.sumatra.isSumatraAvailable
 import javax.swing.JLabel
 import javax.swing.SwingConstants
