+++ src/nl/rubensten/texifyidea/run/LatexCommandLineState.kt
@@ -13,14 +13,19 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiFile
 import nl.rubensten.texifyidea.TeXception
 import nl.rubensten.texifyidea.psi.LatexEnvironment
-import nl.rubensten.texifyidea.run.evince.EvinceConversation
-import nl.rubensten.texifyidea.run.evince.OpenEvinceListener
+import nl.rubensten.texifyidea.run.evince.EvinceForwardSearch
 import nl.rubensten.texifyidea.run.evince.isEvinceAvailable
+import nl.rubensten.texifyidea.run.sumatra.OpenSumatraListener
+import nl.rubensten.texifyidea.run.sumatra.SumatraConversation
+import nl.rubensten.texifyidea.run.sumatra.SumatraForwardSearch
+import nl.rubensten.texifyidea.run.sumatra.isSumatraAvailable
 import nl.rubensten.texifyidea.util.*
 import org.jetbrains.concurrency.runAsync
 import java.io.File
 
 /**
+ * Run the run configuration: start the compile process and initiate forward search (when applicable).
+ *
  * @author Sten Wessel
  */
 open class LatexCommandLineState(environment: ExecutionEnvironment, private val runConfig: LatexRunConfiguration) : CommandLineState(environment) {
@@ -70,78 +75,11 @@ open class LatexCommandLineState(environment: ExecutionEnvironment, private val 
 
         // Open Sumatra after compilation & execute inverse search.
         if (isSumatraAvailable) {
-            handler.addProcessListener(OpenSumatraListener(runConfig))
-
-            // Forward search.
-            run {
-                val psiFile = runConfig.mainFile.psiFile(environment.project) ?: return@run
-                val document = psiFile.document() ?: return@run
-                val editor = psiFile.openedEditor() ?: return@run
-
-                if (document != editor.document) {
-                    return@run
-                }
-
-                // Do not do forward search when editing the preamble.
-                if (psiFile.isRoot()) {
-                    val element = psiFile.findElementAt(editor.caretOffset()) ?: return@run
-                    val environment = element.parentOfType(LatexEnvironment::class) ?: return@run
-                    if (environment.name()?.text != "document") {
-                        return@run
-                    }
-                }
-
-                val line = document.getLineNumber(editor.caretOffset()) + 1
-
-                runAsync {
-                    try {
-                        // Wait for sumatra pdf to start. 1250ms should be plenty.
-                        // Otherwise the person is out of luck ¯\_(ツ)_/¯
-                        Thread.sleep(1250)
-                        SumatraConversation.forwardSearch(sourceFilePath = psiFile.virtualFile.path, line = line)
-                    }
-                    catch (ignored: TeXception) {
-                    }
-                }
-            }
+            SumatraForwardSearch().execute(handler, runConfig, environment)
         }
 
         if(isEvinceAvailable()) {
-            handler.addProcessListener(OpenEvinceListener(runConfig))
-
-            // todo reuse code of above
-            // Forward search.
-            run {
-                val psiFile = runConfig.mainFile.psiFile(environment.project) ?: return@run
-                val document = psiFile.document() ?: return@run
-                val editor = psiFile.openedEditor() ?: return@run
-
-                if (document != editor.document) {
-                    return@run
-                }
-
-                // Do not do forward search when editing the preamble.
-                if (psiFile.isRoot()) {
-                    val element = psiFile.findElementAt(editor.caretOffset()) ?: return@run
-                    val environment = element.parentOfType(LatexEnvironment::class) ?: return@run
-                    if (environment.name()?.text != "document") {
-                        return@run
-                    }
-                }
-
-                val line = document.getLineNumber(editor.caretOffset()) + 1
-
-                runAsync {
-                    try {
-                        // Wait for Evince to start
-                        Thread.sleep(1250)
-                        EvinceConversation.forwardSearch(sourceFilePath = psiFile.virtualFile.path, line = line)
-                    }
-                    catch (ignored: TeXception) {
-                    }
-                }
-            }
-
+            EvinceForwardSearch().execute(runConfig, environment)
         }
 
         return handler
